backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
redis>=5.0,<6.0
greenlet>=3.1,<4.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
pydantic-settings>=2.3,<3.0
python-multipart>=0.0.9,<1.0
